requirements.txt
streamlit
google-generativeai
langchain_community
python-dotenv
langchain_community
pypdf